requirements.txt
python-dotenv
numpy 
pandas
certifi
pymongo
pymongo[srv]==3.12
python-dotenv
scikit-learn
# -e .